requirements.txt
canonicalwebteam.flask-base==2.6.0
canonicalwebteam.discourse==5.4.3
canonicalwebteam.http==1.0.4
canonicalwebteam.blog==6.8.2
canonicalwebteam.image-template==1.3.1
canonicalwebteam.templatefinder==1.0.0
filelock==3.12.0
python-dateutil==2.8.2
pyyaml==6.0
vcrpy-unittest==0.1.7
